// platform/src/payment/payment-method-card-type.ts
export enum PaymentMethodCardType {
    Unknown = 'Unknown',
    Test = 'Test',
    Visa = 'Visa',
    MasterCard = 'MasterCard',
    Discover = 'Discover',
    AmericanExpress = 'AmericanExpress',
    DinersClub = 'DinersClub',
    JCB = 'JCB',
    Switch = 'Switch',
    Solo = 'Solo',
    Dankort = 'Dankort',
    Maestro = 'Maestro',
    Forbrugsforeningen = 'Forbrugsforeningen',
    Laser = 'Laser'
}